+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed: Reduced the Windows install size by about 7 MB, as a regression from last release.
 - Fixed: Logical Pickup placement "All pickups" and expansions with negative amount are now compatible.
 - Fixed: Text such as the Seed Hash in the Async Race window is now selectable via mouse.
+- Removed: The "Consider possible unsafe resources" experimental option has been removed and is now always enabled.
 
 ### Another Metroid 2 Remake